Complete CATCH/THROW, add tests, rework internal error handling
Rather than the system inherited from Beetle of tracking the current error
condition during each instruction and only executing state modifications if
it is currently clear, switch to the Mit system of exiting as soon as an
error occurs. The two systems are intended to be equivalent, but the latter
is easier to get right, though the POP macros can no longer be used as
expressions, which makes the more long-winded to use.

Update the tests and front-end to use the new macros and/or simply use the
underlying functions or data structures directly, which is often shorter.