Add GDB remote protocol support
Run with --gdb to enable connection from gdb; requires gdb built with bee
support.

Reorder tbl_registers.h to a more logical order for presentation of the
registers, rather than alphabetical order, and use it in gdb-stub.c.

Reactivate the assembler test using binutils; make it optional, so that
binutils need not be installed to run the tests.