Add 64-bit support
Conform the logic and memory tests to Mit’s (this was needed because the
previous Bee tests, inherited from Beetle, had 32-bit dependencies).

We no longer need inttypes.h.

Inline stringify.h in cmdline.h, as it is no longer needed for the tests.

One or two other bits of tidy-up.