Add a rule for include/bee/bee.h to src/Makefile.am (fix #7)
To make bee.h be regenerated when bee.h.in is changed by make running in
src/, it is necessary to have a rule for bee.h in src/Makefile.
autoconf/automake can only be persuaded to add a rule to either the
top-level Makefile, or to src/include/bee/Makefile, but the
compiler-generated dependencies, which link the generation of bee.h to its
use by various object files, are only used in src/Makefile.

Hence, add a rule for bee.h to src/Makefile.am similar to that autoconf adds
to the top-level Makefile.