Add stack checks to ensure aborted instructions do nothing
In other words, the intended semantics, and what Mit specifies and does.

Achieve this by adding bee_stack_check() and corresponding CHECKS and CHECKD
macros, and use them for most instructions.

Similarly fix instructions that need special treatment, and change the error
handler so that it does not try to pop an exception frame if the stack could
not possibly hold one.