Fixes to implementation of compact instructions
Have bee_run() use the initial value of ‘ir’. This is necessary to get
single_step() to work correctly with compact instructions, so we can add a
BREAK to ‘ir’ part-way through interpreting an instruction word.

‘ir’ must now always contain a valid instruction word, so ensure that the
inner loop that interprets short instructions maintains this invariant.

Encode NOP as 0. This allows the “natural” initial value of S->ir to work,
now that we no longer read ‘ir’ from ‘pc’ on startup of bee_run().

Also change the interpretation of ‘pc’-relative offsets to be from the value
of ‘pc’ rather than that of the instruction. This simplifies the interpreter
code slightly at the cost of complicating the assembler.

In single_step(), use and update S->ir, so that compact instructions can be
single-stepped correctly, and explicitly skip end-of-word NOP instructions,
to keep from having to worry about this case in the tests.